backend/requirements.txt
aiocache==0.11.1
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.1
fastapi==0.85.0
fastapi-discord==0.2.4
frozenlist==1.3.3
h11==0.14.0
idna==3.4
multidict==6.0.4
pydantic==1.10.9
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.6.3
uvicorn==0.22.0
yarl==1.9.2
mysql-connector-python==8.0.33
